created /azstorage/containers to get list of buckets. Merged all external json into external-buckets.json
