+++ data/external-buckets.json
@@ -1,0 +1,54 @@
+[
+    {
+        "id": "msft",
+        "published": true,
+        "path": "msft/",
+        "url": "https://planetarycomputer.microsoft.com/api/stac/v1/collections",
+        "label": "Microsoft Planetary Computer STAC API",
+        "description": "Searchable spatiotemporal metadata describing Earth science datasets hosted by the Microsoft Planetary Computer",
+        "icon": "fa-brands fa-microsoft",
+        "type": "external",
+        "tags": ["STAC", "API", "Microsoft"],
+        "selected": false
+    },
+    {
+        "id": "msft-mosaic",
+        "published": true,
+        "path": "/stac/v1/collections",
+        "url": "https://planetarycomputer.microsoft.com/api",
+        "label": "Microsoft Planetary Computer STAC API",
+        "description": "Searchable spatiotemporal metadata describing Earth science datasets hosted by the Microsoft Planetary Computer",
+        "icon": "/stac.png",
+        "type": "external",
+        "tags": ["MosaicJSON", "API", "Microsoft"],
+        "selected": false,
+        "isRaster": true,
+        "isStac": true,
+        "isMosaicJSON": true
+    },
+    {
+        "id": "martin",
+        "published": true,
+        "path": "https://martin.undpgeohub.org/index.json",
+        "url": "https://martin.undpgeohub.org",
+        "label": "Martin Vector Tiles API",
+        "description": "Dynamic vector tiles sources from PostGIS database",
+        "icon": "fa-solid fa-database",
+        "type": "external",
+        "tags": ["martin", "API", "UNDP"],
+        "selected": false
+    },
+    {
+        "id": "pgtileserv",
+        "published": true,
+        "path": "https://pgtileserv.undpgeohub.org/index.json",
+        "url": "https://pgtileserv.undpgeohub.org",
+        "label": "pgTileServ Vector Tiles API",
+        "description": "Dynamic vector tiles sources from PostGIS database",
+        "icon": "https://access.crunchydata.com/documentation/pg_tileserv/latest/crunchy-spatial-logo.png",
+        "type": "external",
+        "tags": ["pgtileserv", "API", "UNDP"],
+        "selected": false
+    }
+]
+
